Added IPCStorage and improved api

+++ middlewares/http.js
@@ -1,31 +1,62 @@
 const uuid = require('uuid');
 const url = require('url');
-const rimraf = require('rimraf');
 const { resolve } = require('path');
 const LambdaPool = require('../classes/LambdaPool');
 const RequestEvent = require('../classes/RequestEvent');
 
 let overallLimit = 3000;
 
+const communicationTypes = {
+  ipc: {
+    path: resolve(__dirname, '../classes/IPCStorage'),
+  },
+  file: {
+    path: resolve(__dirname, '../classes/FileStorage')
+  }
+};
+
 /**
- * @param {string} lambdaToInvoke
- * @param {string} handlerKey
- * @param {function} logger
- * @param {string} storageDriverPath
+ * @typedef Communication
+ * @property {string} type
+ * @property {string} [path]
+ */
+
+/**
+ * @typedef Options
+ * @property {string} lambdaPath
+ * @property {string} handlerKey
+ * @property {function} logger
+ * @property {number} limit
+ * @property {Communication} communication
+ */
+
+/**
+ * @param {Options} options
  * @returns {Function}
  */
-const createHttpMiddleware = (lambdaToInvoke, handlerKey = 'handler', logger = () => {}, storageDriverPath, limit) => {
+const createHttpMiddleware = (options = {}) => {
+  const {
+    lambdaPath,
+    handlerKey = 'handler',
+    logger = () => {},
+    limit = overallLimit,
+    communication = {},
+  } = options;
+  const communicationConfig = {
+    ...(communicationTypes[communication.type] && communicationTypes[communication.type]),
+    ...(!communicationTypes[communication.type] && communication),
+    ...(!communication.type && communicationTypes.ipc),
+  };
   // TODO: tmp folders
-  if (!storageDriverPath || typeof storageDriverPath !== "string") return (req, res, next) => { next() };
+  if (!communicationConfig.path || typeof communicationConfig.path !== "string") return (req, res, next) => { next() };
 
   if (limit) {
     overallLimit = limit;
   }
 
-  rimraf.sync(resolve(__dirname, '../requests/*'));
-  rimraf.sync(resolve(__dirname, '../responses/*'));
-  const StorageDriver = require(storageDriverPath);
-  const lambdaPool = new LambdaPool({ storageDriverPath, handlerKey, logger });
+  const StorageDriver = require(communicationConfig.path);
+  if (StorageDriver.start) StorageDriver.start();
+  const lambdaPool = new LambdaPool({ storageDriverPath: communicationConfig.path, handlerKey, logger, communication });
   return (request, response) => {
     const {
       query: queryStringParameters,
@@ -33,34 +64,33 @@ const createHttpMiddleware = (lambdaToInvoke, handlerKey = 'handler', logger = (
     } = url.parse(request.url, true);
 
     /** @var {RequestEvent} event */
-    const event = new RequestEvent;
-    event.httpMethod = request.method.toUpperCase();
-    event.path = path;
-    event.queryStringParameters = queryStringParameters;
-    event.headers = request.headers;
+    const requestEvent = new RequestEvent;
+    requestEvent.httpMethod = request.method.toUpperCase();
+    requestEvent.path = path;
+    requestEvent.queryStringParameters = queryStringParameters;
+    requestEvent.headers = request.headers;
 
     const requestId = uuid.v4();
-    const storage = new StorageDriver(requestId);
+    let storage;
 
     let lambdaInstance;
 
-    logger('Invoking lambda', `${lambdaToInvoke}#${handlerKey}`);
+    logger('Invoking lambda', `${lambdaPath}#${handlerKey}`);
 
     const closeListener = () => {
       if (!response.finished) response.end();
-      storage.destroy();
+      if (storage) storage.destroy();
     };
     Promise.resolve()
-      .then(() => lambdaPool.getLambda(lambdaToInvoke, handlerKey))
+      .then(() => lambdaPool.getLambda(lambdaPath, handlerKey))
       .then(instance => {
         instance.addEventListenerOnce('close', closeListener);
 
         lambdaInstance = instance;
-        return Promise.resolve();
+        storage = new StorageDriver(requestId, instance);
+        return Promise.resolve(instance);
       })
-      .then(() => storage.setRequest(event))
-      .then(() => new Promise(res =>  lambdaInstance.invoke(requestId, res))
-      .then(() => storage.getResponse()))
+      .then(() => new Promise(res => lambdaInstance.invoke(requestId, requestEvent, res)))
       /** @var {ResponseEvent} responseEvent */
       .then(responseEvent => {
         if (responseEvent.statusCode) {
@@ -74,7 +104,7 @@ const createHttpMiddleware = (lambdaToInvoke, handlerKey = 'handler', logger = (
         }
 
         lambdaInstance.removeEventListener('close', closeListener);
-        return storage.destroy();
+        return storage ? storage.destroy() : Promise.resolve();
       })
       .catch(err => {
         logger(err);
